refactor: redis config loader
